QOL Fixes
The close button for the Pokemon stat screen is now sticky so you no longer need to scroll down to close the view. The gender/shiny icons for Pokemon no longer cause the screen to shake when switching. You may need to re-add the app to your home screen to see changes. Please note these changes are currently being tested in the web app and will be rolled out for Android native once the next major update rolls out (soon!)